% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_grid_sp_beta.R
\docType{data}
\name{data.used}
\alias{data.used}
\title{An example dataset}
\format{A data frame with 932 rows and 3 variables}
\usage{
data(data.used)
}
\description{
A dataframe containing multiple taxa occurrences
}
\keyword{datasets}
