% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_grid_sp_beta.R
\name{grid_couting_tax}
\alias{grid_couting_tax}
\title{Counting and rasterize multiple taxa occurrences}
\usage{
grid_couting_tax(data, long, lat, sp, resolution = 10)
}
\arguments{
\item{data}{A dataframe with taxa field, longitude and latitude}

\item{long}{A character string indicating the column name containing longitude in decimal degrees}

\item{lat}{A character string indicating the column name containing latitude in decimal degrees}

\item{sp}{A character string indicating the column name of taxa}

\item{resolution}{A numeric indicating the resolution in kilometres of the grid}
}
\description{
Defining a grid of given resolution and counting occurrences for multiple taxa
}
